refactor: updates class export
